qemu-ga: partial fix for fsfreeze, and a bunch of junk
when we last left off we'd realized we can't use a persistant state file
due to atime updates blocking during fsfreeze. lulz.

the only solution other than requiring noatime (or playing dangerous
games with relatime semantics) is to use file *existance* as the flag
for fsfreeze, since open()'s dont cause the atime update.

that should do it.

for filehandle offsets, that's not feasible, and guaranteeing persistant
across reboots (a requirement for unique, sequenced IDs) is difficult
since common directories like /tmp and /var/run are, at least on ubuntu,
tmpfs mounts.

one solution is to reserve the upper 31 (32 if we allow negative
filehandle IDs) bits for a timestamp. For 1 second resolution it means
we're fine unless they manage to create 2^32 FDs before qemu-ga restarts
(assuming we sleep(min_interval) to make sure the timestamp updates).

We can also use higher-resolution timestamps, and reduce the number of
bits for the sequential IDs. This still works since they have less time
to overflow the upper 31-bits before we bump up the clock offset.