Makefile: add infrastructure to incorporate qidl-generated files
This will rebuild the qidl-generated qapi schemas in
$(SRC_DIR)/qidl-generated whenever we do a build where of the annotated
sources has been modified. As a result, changes will be reflected in the
working tree so we can detect changes and commit them in cases where the
changes are correct/expected.

We also add the necessary targets for compiling and linking in the
qapi-generated visitors for device state.

RTC is the first user.