qapi: qapi.py, make json parser more robust
Currently the QAPI JSON parser expects a very particular style of code
indentation, the major one being that terminating curly/square brackets are
not on placed on a seperate line. This is incompatible with most
pretty-print formats, so make it a little more robust by supporting
these cases.

Also add support for parsing numerical fields. Currently they are
ignored.

QIDL will make use of both of these changes with the schemas it
generates.

Reviewed-by: Paolo Bonzini <pbonzini@redhat.com>