qom-fuse: force single-threaded mode to avoid QMP races
python-fuse defaults to multi-threaded handling of filesystem calls.
In the case of QOM this can lead to threads reading QMP responses to
requests executed by other threads, causing all sorts of strangeness
when interacting with QOM mounts. For instance:

  mdroth@loki:~/w/qom$ ls -l machine | grep type
  ls: error while loading shared libraries: tls/libacl.so.1: \
  cannot read file data: Error 21

We fix this by unconditionally passing in the -s option, which forces
single-threaded/serialized execution of filesystem calls. This flag is
only honored if we pass dash_s_do='setsingle' to the Fuse class
constructor and use the parse() method of that class to pass in
arguments, so we do that here as well.