qemu-ga: add guest-file-* interfaces for w32
The FILE-based functions we use for the POSIX implementation don't quite
work since O_NONBLOCK is not supported for MinGW builds. This is okay
for most things, but causes us to block indefinitely for pipes, which
means we can't do fun things like interact with stdio from processes we
launch via the upcoming guest-exec-async commands. This is why these
interfaces weren't enabled for w32 originally as well.

GIOChannels offer non-blocking I/O however, so that makes it good enough
for basic file access and interactive commands. It does not however offer
a means to implement guest-file-seek...we can seek, but we can't, afaik,
return the current position as is expected by guest-file-seek. This why
we only use this specifically for w32. If we figure this out in the
future however it may be worth consolidating around GIOChannel.

Signed-off-by: Michael Roth <mdroth@linux.vnet.ibm.com>