*spapr_pci: rework spapr_map_BARs to use new mr func
This re-works the code a little to remove reliance on
MemoryRegionSections and instead just return a hwaddr based on the
results of the search.

We also add logic to search in increments of r->size, since the
pci code expects that and masks of any that aren't r->size-aligned.

Also, pci code has some checks against PCI_COMMAND region memory
it'll allow BARs to be programmed, so add those calls as well.

Signed-off-by: Michael Roth <mdroth@linux.vnet.ibm.com>