*spapr_pci: add missing OF properties for PCI nodes
The main ones we were missing to allow discovery by kernel was
name, reg, and assigned-addresses. With these in place the device
will *in some cases* get added to /sys/devices, but there are
page fault errors present that need to be worked through. Possibly
we need to begin region allocation at higher addresses for io/mmio
regions, as that seems to be the only difference between the
hotplug assignments vs. SLOF

Signed-off-by: Michael Roth <mdroth@linux.vnet.ibm.com>